@skyzh 2023-08-11
it seems that TPC-H plans are slower to generate than before, but we can improve it later. (make `apply_planner_test`)