@wangrunji0408 2023-08-11
Yes. The current CBO has not been fine-tuned. It can be very slow especially with many table joins. We should rethink the strategy to apply rules and may need to explore how to perform RBO in the egg framework. 🥵